modify read_cmd, first enter error
