+++ Makefile
@@ -2,15 +2,22 @@ NAME = minishell
 CC = gcc
 # CFLAGS = -Wall -Wextra -Werror
 SRCS_DIR =	./srcs	\
-			./srcs/gnl
+			./srcs/utils
 
 OBJS_DIR = ./objs
 INC_DIR = ./includes
 
 SRCS =	./srcs/main.c	\
-		./srcs/tokenizer.c	\
-		./srcs/gnl/get_next_line_bonus.c	\
-		./srcs/gnl/get_next_line_utils_bonus.c
+		./srcs/separate.c	\
+		./srcs/read_cmd.c	\
+		./srcs/utils/ft_str_char_join.c	\
+		./srcs/utils/ft_strlen.c	\
+		./srcs/utils/ft_strchr.c	\
+		./srcs/utils/ft_strdup.c	\
+		./srcs/utils/ft_strtrim.c	\
+		./srcs/utils/ft_substr.c	\
+		./srcs/utils/ft_memmove.c	\
+		# ./srcs/tokenizer_struct.c
 
 OBJS = $(addprefix $(OBJS_DIR)/, $(notdir $(SRCS:.c=.o)))
 vpath %.c $(SRCS_DIR)
