+++ includes/minishell.h
@@ -3,8 +3,12 @@
 
 # include <stdio.h>
 # include <unistd.h>
-# include <get_next_line_bonus.h>
-# include <structures.h>
+# include <termios.h>
+# include <stdlib.h>
+# include <signal.h>
+# include "structures.h"
+# include "read_cmd.h"
+# include "utils.h"
 
 # define OFF 0
 # define ON 1
