+++ includes/utils.h
@@ -1,11 +1,12 @@
 #ifndef UTILS_H
 # define UTILS_H
 
-int		ft_strlen(char *s);
+int		ft_strlen(const char *s);
 char	*ft_str_char_join(char *before, char c);
 char	*ft_strtrim(char const *s1, char const *set);
 char	*ft_substr(char const *s, unsigned int start, size_t len);
 char	*ft_strdup(const char *s);
 char	*ft_strchr(const char *s, int c);
+void	*ft_memmove(void *dest, const void *src, size_t n);
 
 #endif
