+++ srcs/utils/ft_strlen.c
@@ -1,6 +1,6 @@
 #include "minishell.h"
 
-int		ft_strlen(char *s)
+int		ft_strlen(const char *s)
 {
 	int	i;
